[firebase-cleanup] separates firebase configuration from push notification logic (import fixes on rebase)
