+++ api/survey_api.py
@@ -4,8 +4,9 @@ from authentication.admin_authentication import authenticate_researcher_study_ac
 from database.schedule_models import AbsoluteSchedule, RelativeSchedule, WeeklySchedule
 from database.survey_models import Survey
 from libs.json_logic import do_validate_survey
-from libs.push_notification_config import (repopulate_absolute_survey_schedule_events,
+from libs.push_notification_helpers import (repopulate_absolute_survey_schedule_events,
     repopulate_relative_survey_schedule_events, repopulate_weekly_survey_schedule_events)
+
 
 survey_api = Blueprint('survey_api', __name__)
